revert unnecessary use of int64_t types in blockmap handling
# Conflicts:
#	src/doom/p_setup.c

+++ src/doom/p_setup.c
@@ -86,9 +86,9 @@ static int      totallines;
 // Blockmap size.
 int		bmapwidth;
 int		bmapheight;	// size in mapblocks
-int64_t*	blockmap;	// int for larger maps // [crispy] BLOCKMAP limit
+int32_t*	blockmap;	// int for larger maps // [crispy] BLOCKMAP limit
 // offsets in blockmap are from here
-int64_t*	blockmaplump; // [crispy] BLOCKMAP limit
+int32_t*	blockmaplump; // [crispy] BLOCKMAP limit
 // origin of block map
 fixed_t		bmaporgx;
 fixed_t		bmaporgy;
@@ -1260,15 +1260,15 @@ void P_LoadBlockMap (int lump)
 
     blockmaplump[0] = SHORT(wadblockmaplump[0]);
     blockmaplump[1] = SHORT(wadblockmaplump[1]);
-    blockmaplump[2] = (int64_t)(SHORT(wadblockmaplump[2])) & 0xffff;
-    blockmaplump[3] = (int64_t)(SHORT(wadblockmaplump[3])) & 0xffff;
+    blockmaplump[2] = (int32_t)(SHORT(wadblockmaplump[2])) & 0xffff;
+    blockmaplump[3] = (int32_t)(SHORT(wadblockmaplump[3])) & 0xffff;
 
     // Swap all short integers to native byte ordering.
   
     for (i=4; i<count; i++)
     {
 	short t = SHORT(wadblockmaplump[i]);
-	blockmaplump[i] = (t == -1) ? -1l : (int64_t) t & 0xffff;
+	blockmaplump[i] = (t == -1) ? -1l : (int32_t) t & 0xffff;
     }
 
     Z_Free(wadblockmaplump);
